refactor: move cart methods to cart servce
