+++ e-commerce/src/app/store/effects.ts
@@ -127,13 +127,13 @@ export default class EcommerceEffects {
         this.store.pipe(select(selectAccessToken)), // Select access token from store
       ),
       switchMap(([, accessToken]) =>
-        this.ecommerceApiService.getUserCart(accessToken).pipe(
+        this.cartService.getUserCart(accessToken).pipe(
           switchMap((response) => {
             if (response.status === 200) {
               const cart = response.body; // Assuming response.body contains the body of the response
               return of(actions.loadUserCartSuccess({ cartBase: cart! }));
             }
-            return this.ecommerceApiService.createUserCart(accessToken).pipe(
+            return this.cartService.createUserCart(accessToken).pipe(
               map((cartBase) => actions.loadUserCartSuccess({ cartBase })),
               catchError((error) => of(actions.loadUserCartFailure({ error }))),
             );
